requirements.txt
# Core requirements
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
mypy==1.5.1
tqdm==4.66.1
python-dotenv==1.0.0
sqlalchemy==2.0.20
pydantic==2.3.0
click==8.1.7 